// src/types/api.ts
export interface UserSafeObject {
  id: string
  username: string
  email: string
  role: string
  designation?: string
  avatarUrl?: string
  about?: string
  isActive: boolean
  createdAt: string
  updatedAt: string
}
// ...
export interface LoginResponse {
  success: boolean
  message: string
  token?: string
  user?: UserSafeObject
}

export interface SignupRequest {
  user_name: string
  email: string
  password: string
  user_level?: 'participant'
}

export interface SignupResponse {
  success: boolean
  message: string
  user?: UserSafeObject
  token?: string
}